MAINT: use _asarray_validated instead of ad-hoc validation
