+++ scipy/linalg/decomp.py
@@ -23,64 +23,13 @@ from numpy import array, asarray_chkfinite, asarray, diag, zeros, ones, \
         isfinite, inexact, nonzero, iscomplexobj, cast, flatnonzero, conj
 # Local imports
 from scipy._lib.six import xrange
+from scipy._lib._util import _asarray_validated
 from .misc import LinAlgError, _datacopied, norm
 from .lapack import get_lapack_funcs
 from .blas import get_blas_funcs
 
 
 _I = cast['F'](1j)
-
-
-def _asarray_validated(a, check_finite=True,
-                       sparse_ok=False, objects_ok=False, mask_ok=False):
-    """
-    Helper function for scipy argument validation.
-
-    Many scipy linear algebra functions do support arbitrary array-like
-    input arguments.  Examples of commonly unsupported inputs include
-    matrices containing inf/nan, sparse matrix representations, and
-    matrices with complicated elements.
-
-    Parameters
-    ----------
-    a : array_like
-        The array-like input.
-    check_finite : bool, optional
-        Whether to check that the input matrices contain only finite numbers.
-        Disabling may give a performance gain, but may result in problems
-        (crashes, non-termination) if the inputs do contain infinities or NaNs.
-        Default: True
-    sparse_ok : bool, optional
-        True if scipy sparse matrices are allowed.
-    objects_ok : bool, optional
-        True if arrays with dype('O') are allowed.
-    mask_ok : bool, optional
-        True if masked arrays are allowed.
-
-    Returns
-    -------
-    ret : ndarray
-        The converted validated array.
-
-    """
-    if not sparse_ok:
-        import scipy.sparse
-        if scipy.sparse.issparse(a):
-            msg = ('Sparse matrices are not supported by this function. '
-                   'Perhaps one of the scipy.linalg.sparse functions '
-                   'would work instead.')
-            raise ValueError(msg)
-    if not mask_ok:
-        if np.ma.isMaskedArray(a):
-            raise ValueError('masked arrays are not supported')
-    if check_finite:
-        a = np.asarray_chkfinite(a)
-    else:
-        a = np.asarray(a)
-    if not objects_ok:
-        if a.dtype is np.dtype('O'):
-            raise ValueError('object arrays are not supported')
-    return a
 
 
 def _make_complex_eigvecs(w, vin, dtype):
